backend/pyproject.toml
[project]
name = "test-bmad-backend"
version = "0.1.0"
description = "FastAPI backend for test-bmad."
requires-python = ">=3.12"
dependencies = [
  "google-cloud-vision>=3.7,<4.0",
  "fastapi==0.129.0",
  "langchain-core==0.3.61",
  "pillow==11.2.1",
  "pydantic==2.11.9",
  "pypinyin==0.55.0",
  "python-multipart==0.0.20",
  "uvicorn[standard]==0.37.0",
]

[dependency-groups]
dev = [
  "boto3==1.37.26",
  "httpx==0.28.1",
  "pytest==8.4.2",
  "ruff==0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N"]

[tool.pytest.ini_options]
pythonpath = [".", "tests"]
testpaths = ["tests"]
